landing page 및 root page(home) 작업
