+++ components/user/UserBottomNav.vue
@@ -24,24 +24,6 @@
       </span>
       <span class="text-xs font-medium text-gray-600">홈</span>
     </NuxtLink>
-    <NuxtLink to="/stores" class="flex flex-col items-center">
-      <span class="p-2 mb-1">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          stroke-width="1.5"
-          stroke-linecap="round"
-          stroke-linejoin="round"
-          class="lucide lucide-search w-6 h-6"
-        >
-          <circle cx="11" cy="11" r="8" />
-          <path d="m21 21-4.3-4.3" />
-        </svg>
-      </span>
-      <span class="text-xs font-medium text-gray-600">검색</span>
-    </NuxtLink>
     <NuxtLink to="/my/tickets" class="flex flex-col items-center">
       <span class="p-2 mb-1">
         <svg
