+++ components/user/stores/detail/StoreDetailImages.vue
@@ -1,5 +1,20 @@
 <template>
-  <div class="relative w-full aspect-video bg-black">
+  <div class="relative w-full aspect-square bg-slate-100">
+    <button type="button" class="z-50 ml-4 mt-2 absolute" @click="goBack">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        fill="none"
+        stroke="currentColor"
+        stroke-width="1.5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        class="lucide lucide-arrow-left w-7"
+      >
+        <path d="m12 19-7-7 7-7" />
+        <path d="M19 12H5" />
+      </svg>
+    </button>
     <img
       :src="currentImage"
       :alt="restaurantName"
@@ -75,6 +90,10 @@
 
 <script setup>
 import { ref, computed } from 'vue';
+
+const goBack = () => {
+  navigateTo('/');
+};
 
 const props = defineProps({
   images: {
